+++ apps/app/capacitor.config.ts
@@ -8,6 +8,9 @@ const config: CapacitorConfig = {
     androidScheme: 'https',
   },
   plugins: {
+    CapacitorHttp: {
+      enabled: true,
+    },
     SystemBars: {
       insetsHandling: 'css',
       style: 'DEFAULT',